+++ src/main/java/com/example/demo/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package com.example.demo.repository;
 
-import com.example.demo.dao.User;
+import com.example.demo.entity.User;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
@@ -8,11 +8,9 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
-
 
 @Repository
-public interface UserRepository extends CrudRepository<User, Integer> {
+public interface UserRepository extends CrudRepository<User, Long> {
 
     Iterable<User> findAll(Pageable pageable);
     @Modifying(clearAutomatically = true)
